pwa/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
pyjwt==2.8.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
